chore: move and rename db types
