+++ types/custom.ts
@@ -1,40 +1,9 @@
+import { outputTransactionFields } from '@/utils/helpers';
 import { BarChartProps, Color } from '@tremor/react';
+import type { DateRange } from 'react-day-picker';
 import { components } from './up-api';
 
-import { outputTransactionFields } from '@/utils/helpers';
-import type { Binary } from 'bson';
-import type { DateRange } from 'react-day-picker';
-
-interface TransactionResourceAttributes
-  extends Omit<
-    components['schemas']['TransactionResource']['attributes'],
-    'createdAt' | 'settledAt'
-  > {
-  createdAt: Date;
-  settledAt: Date | null;
-}
-
-// Account info as stored in db
-export interface AccountResource
-  extends Omit<components['schemas']['AccountResource'], 'id'> {
-  _id: string;
-}
-
 export type AccountType = components['schemas']['AccountTypeEnum'];
-
-export interface CategoryResource
-  extends Omit<components['schemas']['CategoryResource'], 'id'> {
-  _id: string;
-}
-
-export interface DbTransactionResource
-  extends Omit<
-    components['schemas']['TransactionResource'],
-    'id' | 'attributes'
-  > {
-  _id: Binary;
-  attributes: TransactionResourceAttributes;
-}
 
 export type SerialisedDbTransactionResource = ReturnType<
   typeof outputTransactionFields
